# routes/site.js
const express = require('express');
const _ = require('lodash');

const { authenticateMiddleware } = require('../middleware');

const {
  getUserStatistics,
} = require('../tools/site');

const router = express.Router();

router.get('/api/site/getUserStatistics', authenticateMiddleware, async (req, res) => {
  /*
    #swagger.summary = 'Get user statistics'

    #swagger.description = 'Get user statistics'

    #swagger.parameters['Authorization'] = {
      in: 'header',
      description: 'Example: Bearer xxxxxx',
      required: 'true',
      schema: {
        Authorization: 'Bearer ',
      },
    }

    #swagger.parameters['todayStart'] = {
      in: 'query',
      description: 'Number of milliseconds, today startOf timestamp',
      schema: {
        todayStart: 1666368000000,
      },
    }

    #swagger.responses[200] = {
      description: 'Get user statistics successfully.',
      schema: {
        message: 'Get user statistics successfully.',
        userStatistics: {
          userTotal: 3,
          todayActiveUserTotal: 1,
          lastSevenDayActiveUserAverage: 0,
        },
      },
    }
  */

  if (!req.user) {
    return res.status(403).json({
      message: 'Forbidden.',
    });
  }

  let todayStart;
  if (!_.isNaN(parseInt(req.query.todayStart, 10))) {
    todayStart = parseInt(req.query.todayStart, 10);
  }

  const userStatistics = await getUserStatistics({ todayStart });

  res.status(200).json({
    message: 'Get user statistics successfully.',
    userStatistics,
  });
});

module.exports = router;