# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "aaa@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "qwerQWER1@"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Sign in suceesfully."
                },
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "token": {
                  "type": "string",
                  "example": "xxxxxx"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "description": "Register",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "ccc@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "qwerQWER1@"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Register successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Register suceesfully."
                },
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "token": {
                  "type": "string",
                  "example": "xxxxxx"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/auth/refreshToken": {
      "post": {
        "summary": "Refresh token",
        "description": "Refresh token",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer xxxxxx",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string",
                  "example": "Bearer "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh token successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Refresh token suceesfully."
                },
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "token": {
                  "type": "string",
                  "example": "xxxxxx"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout",
        "description": "Logout, clear cookie token, and redirect /sign-in",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get user list",
        "description": "Get user list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page, default 1",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "perPage, default 10",
            "schema": {
              "type": "object",
              "properties": {
                "perPage": {
                  "type": "number",
                  "example": 10
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer xxxxxx",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string",
                  "example": "Bearer "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get user list successfully.  \n LastSession and signUpAt the number of milliseconds.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Get user list successfully."
                },
                "users": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "numberOfLogon": {
                        "type": "number",
                        "example": 1
                      },
                      "lastSession": {
                        "type": "number",
                        "example": 1666491318365
                      },
                      "signUpAt": {
                        "type": "number",
                        "example": 1666491318365
                      }
                    }
                  }
                },
                "dataTotal": {
                  "type": "number",
                  "example": 3
                },
                "page": {
                  "type": "number",
                  "example": 1
                },
                "perPage": {
                  "type": "number",
                  "example": 10
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/users/{id}/updateName": {
      "post": {
        "summary": "Update user's name",
        "description": "Update name, only your own name can be updated",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer xxxxxx",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string",
                  "example": "Bearer "
                }
              }
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Jack"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update name successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Update name suceesfully."
                },
                "userId": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/site/getUserStatistics": {
      "get": {
        "summary": "Get user statistics",
        "description": "Get user statistics",
        "parameters": [
          {
            "name": "todayStart",
            "in": "query",
            "description": "Number of milliseconds, today startOf timestamp",
            "schema": {
              "type": "object",
              "properties": {
                "todayStart": {
                  "type": "number",
                  "example": 1666368000000
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer xxxxxx",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string",
                  "example": "Bearer "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get user statistics successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Get user statistics successfully."
                },
                "userStatistics": {
                  "type": "object",
                  "properties": {
                    "userTotal": {
                      "type": "number",
                      "example": 3
                    },
                    "todayActiveUserTotal": {
                      "type": "number",
                      "example": 1
                    },
                    "lastSevenDayActiveUserAverage": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}